refactored lib/main and put all the code regarding the scrollable_list in its own file
