+++ lib/main.dart
@@ -1,6 +1,9 @@
 import 'dart:math';
 import 'package:flutter/material.dart';
 import 'package:finesse_nation/addEvent.dart';
+import 'package:finesse_nation/widgets/buildFinesseCard.dart';
+import 'package:finesse_nation/Finesse.dart';
+import 'package:finesse_nation/widgets/buildFinesseList.dart';
 
 void main() => runApp(MyApp());
 
@@ -22,7 +25,7 @@ class MyApp extends StatelessWidget {
         // is not restarted.
         primarySwatch: Colors.blue,
       ),
-      home: MyHomePage(title: 'Rahbert'),
+      home: MyHomePage(title: 'Finesse Nation'),
     );
   }
 }
@@ -46,25 +49,6 @@ class MyHomePage extends StatefulWidget {
 }
 
 class _MyHomePageState extends State<MyHomePage> {
-  int _counter = 0;
-  double _scale = 1;
-
-//  Color _color = Color.fromRGBO(0, 255, 0, 0.1);
-  void _incrementCounter() {
-    var rng = new Random();
-    setState(() {
-      // This call to setState tells the Flutter framework that something has
-      // changed in this State, which causes it to rerun theflect the updated valuese build method below
-      // so that the display can r. If we changed
-      // _counter without calling setState(), then the build method would not be
-      // called again, and so nothing would appear to happen.
-      _counter++;
-      var old = _scale;
-      while (_scale == old) {
-        _scale = rng.nextInt(9) + 1.0; //rng.nextDouble() * 9 + 1;
-      }
-    });
-  }
 
   @override
   Widget build(BuildContext context) {
@@ -91,97 +75,15 @@ class _MyHomePageState extends State<MyHomePage> {
         icon: Icon(Icons.add),
         backgroundColor: Colors.pink,
       ),
-      body: _buildFinesseList(),
-
+      body: buildFinesseList(),
       // This trailing comma makes auto-formatting nicer for build methods.
-      );
-    }
-
-    List<Finesse> _finesses = new List();
-    getFinesses() {
-      _finesses.add(Finesse("Fin 1 Title", "Fin 1 Duration", "images/rahbert.png"));
-      _finesses.add(Finesse("Fin 2 Title", "Fin 2 Duration", "images/bemzo.jpg"));
-      _finesses.add(Finesse("Fin 3 Title", "Fin 3 Duration", "images/rahbert.png"));
-      _finesses.add(Finesse("Fin 4 Title", "Fin 4 Duration", "images/bemzo.jpg"));
-      _finesses.add(Finesse("Fin 5 Title", "Fin 5 Duration", "images/rahbert.png"));
-      _finesses.add(Finesse("Fin 6 Title", "Fin 6 Duration", "images/bemzo.jpg"));
-      _finesses.add(Finesse("Fin 7 Title", "Fin 7 Duration", "images/rahbert.png"));
-      _finesses.add(Finesse("Fin 8 Title", "Fin 8 Duration", "images/bemzo.jpg"));
-      _finesses.add(Finesse("Fin 9 Title", "Fin 9 Duration", "images/rahbert.png"));
-      _finesses.add(Finesse("Fin 10 Title", "Fin 10 Duration", "images/bemzo.jpg"));
-      _finesses.add(Finesse("Fin 11 Title", "Fin 11 Duration", "images/rahbert.png"));
-    }
-
-
-    Widget _buildFinesseList() {
-      return new Container(
-        decoration: BoxDecoration(
-          gradient: LinearGradient(
-            begin: Alignment.topLeft,
-            end: Alignment.topRight,
-            colors: [Colors.lightBlue, Colors.lightGreen],
-          )
-        ),
-        child: new Center(
-
-          child:
-            ListView.builder(
-              itemBuilder: (context, i) {
-                if(i == 0){
-                  getFinesses();
-                }
-                if (i.isOdd) return Divider();
-                final index = i ~/ 2;
-                if(index >= _finesses.length){
-                  _finesses.addAll(_finesses.take(10));
-                }
-                return _buildFinesseCard(_finesses[index]);
-              })
-        )
-      );
-    }
-
-
-  Widget _buildFinesseCard(Finesse fin) {
-    return Card(
-      color: Colors.lime[200],
-      child: Column(
-          mainAxisSize: MainAxisSize.min,
-          children: <Widget>[
-             Image.asset(fin.getImage()),
-             ListTile(
-              leading: Icon(Icons.accessible_forward),
-              title: Text(fin.getTitle()),
-              subtitle: Text(fin.getBody()),
-            ),
-          ]),
     );
   }
+
+
+
 }
 
 
-class Finesse{
-  String title;
-  String body;
-  String duration;
-  String image;
 
-  Finesse(String title, String body, String img){
-    this.title = title;
-    this.body = body;
-    this.image = img;
-  }
-
-  String getImage(){
-    return image;
-  }
-
-  String getBody(){
-    return body;
-  }
-
-  String getTitle(){
-    return title;
-  }
-}
